requirements.txt
certifi==2020.12.5
chardet==4.0.0
grakn-client==2.0.0a7
grakn-protocol==2.0.0a9
graknprotocol==2.0.0a5
grpcio==1.35.0
idna==2.10
numpy==1.20.1
pandas==1.2.2
protobuf==3.14.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
untangle==1.1.1
urllib3==1.26.3
wget==3.2
